bgp-analyzer/requirements.txt
pybgpstream
prometheus_client
python-dotenv
